requirements.txt
asgiref==3.9.1
Django==5.2.5
django-extensions==3.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
mkdocs==1.4.2
mkdocs-material==9.1.12
gunicorn==23.0.0
dj-database-url==3.0.1
psycopg[binary]==3.2.9
python-dotenv==1.0.0
whitenoise==6.5.0
transitions
django-jazzmin
django-autocomplete-light
django-import-export